% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power regression.R
\name{plot.simPower.moderation}
\alias{plot.simPower.moderation}
\title{Plot method for simPower.moderation}
\usage{
\method{plot}{simPower.moderation}(x, pval = 0.05, ...)
}
\arguments{
\item{x}{an object of class simPower.moderation}
}
\description{
This function allows you to plot the power of an object of class simPower.moderation
}
\examples{
simPower.moderation()
}
\keyword{interaction}
\keyword{moderation}
\keyword{plot}
\keyword{power}
\keyword{regression}
