% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrPwrSim3.R
\name{regrPwrSim}
\alias{regrPwrSim}
\title{Simulation for power moderation function for one sample size}
\usage{
regrPwrSim(n = 100, predictors = NULL, cor = NULL, betas = NULL,
  interactions = NULL, sig.level = 0.05, maxiter = 1000,
  monteCarlo = FALSE, digits = 2)
}
\arguments{
\item{predictors}{A number of predictor variables}

\item{cor}{A vector or matrix containing the correlation bewtween the predictors, if NULL random correlations will be used}

\item{betas}{A vector regression coeffcients (effect sizes)}

\item{interactions}{A list of vectors containing a pair of variables}

\item{sig.level}{Type I error level for power analysis}
}
\description{
This function allows you to estimate the power in an linear model with interaction effects.
}
\examples{
regrPwrSim(n=100, predictors = 3, betas = c(.4,.3.,.2), maxiter= 1000)
}
\keyword{interaction}
\keyword{moderation}
\keyword{power}
\keyword{regression}
